Chapter-Docker/my_first_consumer/requirements.txt
websockets==12.0
requests==2.31.0
python-dotenv==1.1.1
colorama==0.4.6 